refactor: move core modules into package
