+++ src/mr_manager/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core package for configuration, cache, and repository discovery logic."""
