+++ src/mr_manager/core/cache.py
