+++ src/mr_manager/core/discovery.py
